refactor(registry): remove eservicetoken enum and update registry interface
BREAKING CHANGE: - Remove EServiceToken enum previously used for common service registration
- Change Registry interface to use getName() method instead of name property
- Update documentation examples to use custom tokens

Replace predefined tokens with user-defined tokens. Update Registry interface to require getName()
method instead of name property. Clean up imports throughout the codebase for better organization.
