+++ docs/api-reference/enums/page.mdx
@@ -18,20 +18,3 @@ export enum ELoggerLogLevel {
 
 - **Hierarchy**: `TRACE` > `DEBUG` > `INFO` > `WARN` > `ERROR` (most verbose to least verbose).
 - **Usage**: Used in [`IConsoleLoggerOptions`](/docs/cladi/api-reference/interfaces#iconsoleloggeroptions) to set the minimum logging level.
-
-## `EServiceToken`
-
-*Defined in: `src/domain/enum/service-token.enum.ts`*
-
-Defines standard `symbol` tokens for commonly used services within the [`IContainer`](/docs/cladi/api-reference/interfaces#icontainer).
-
-```typescript filename="src/domain/enum/service-token.enum.ts" copy
-export enum EServiceToken {
-	CONTAINER = "container", // Typically used to register the container itself
-	FACTORY = "factory",     // For the main application Factory
-	LOGGER = "logger",       // For the main application Logger
-	REGISTRY = "registry",   // For the main application Registry
-}
-```
-
-- **Purpose**: Provides consistent, well-known symbols for core service registration and retrieval, reducing the need for custom symbols for basic services.
