+++ docs/services/logging/page.mdx
@@ -128,4 +128,4 @@ const debugLogger = createLogger({
 
 ## Integrating with Dependency Injection
 
-Typically, you create a logger instance early in your application setup and register it with the [`Container`](/docs/cladi/core-concepts/container) using `EServiceToken.LOGGER`. Other services can then receive the logger via constructor injection.
+Typically, you create a logger instance early in your application setup and register it with the [`Container`](/docs/cladi/core-concepts/container). Other services can then receive the logger via constructor injection.
