+++ src/domain/enum/index.ts
@@ -1,2 +1,1 @@
 export { ELoggerLogLevel } from "./logger-log-level.enum";
-export { EServiceToken } from "./service-token.enum";
