--- src/domain/enum/service-token.enum.ts
@@ -1,9 +1,0 @@
-/**
- * Service tokens for dependency injection.
- */
-export enum EServiceToken {
-	CONTAINER = "container",
-	FACTORY = "factory",
-	LOGGER = "logger",
-	REGISTRY = "registry",
-}
